Respect badge position when tab is not focused.
Fixes #91

Prior to this change, with `position: 'up'`, the badge would still appear at `position: 'bottom'` when the tab wasn't focused. When the tab isn't focused (ie the document is hidden), we use the special `'none'` animation type to place the badge. However, this animation type wasn't getting transformed for the `'up'` and `'left'` positions -- only the current animation type was. Now we transform all the animations. This fixes the issue.